Improve clarity of the handle example
I have rebased the original PR on top of c05d3e0 -- the commit that
moved the Handle mechanics back into the Freshness section.

Further to the rebase, I have applied the following changes:
* The intro text is removed because it has been made redundant;
* HT has been removed from the sequence diagram as well as from text.

This should be now in a state from where Henk and Eric can bring it to
completion.

Fixes #198