More fixes to handle example
Biggest bug in the last pull request is that the diagram had
the attester sending a different Attestation Result than the
one the Verifier gave it.

Second issue was that the lines didn't line up.

Third issue was that it was missing the tie to the "epoch window"
discussion in epochfreshness to explain how the Relying Party deals
with the mismatched handles that the example diagram has.

Fixes #198

Signed-off-by: Dave Thaler <dthaler@ntdev.microsoft.com>